# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscribeRequest {
  email: string;
  name?: string;
}

export interface Subscriber {
  id: number;
  email: string;
  name?: string | null;
  subscribedAt: string;
  isActive: boolean;
}

export interface SubscribeResponse {
  success: boolean;
  message: string;
  subscriber?: Subscriber;
}

export interface UnsubscribeRequest {
  email: string;
}

export interface MessageResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SubscribersResponse {
  subscribers: Subscriber[];
  total: number;
}

export interface EventItem {
  title: string;
  date: string;
  venue: string;
  description: string;
  link?: string | null;
  category: string;
  imageUrl?: string | null;
  source?: string | null;
  featured?: boolean;
}

export interface Digest {
  id: number;
  weekOf: string;
  subject: string;
  intro: string;
  events: EventItem[];
  sentAt?: string | null;
  sentCount: number;
  createdAt: string;
}

export interface DigestResponse {
  digest: Digest;
}

export interface DigestListResponse {
  digests: Digest[];
}

export interface GenerateDigestRequest {
  /** ISO date string for the week (defaults to next Sunday) */
  weekOf?: string;
  /** Optional personal notes to include in the intro */
  customNotes?: string;
}

export interface SendDigestRequest {
  digestId: number;
  /** If provided, only sends to this test email */
  testEmail?: string;
}

export interface TenantPublic {
  slug: string;
  name: string;
  city: string;
  accentColor: string;
  categories: string[];
}

export interface TenantConfigResponse {
  tenant: TenantPublic;
}

export interface TenantsListResponse {
  tenants: TenantPublic[];
}

export interface LeaderboardEntry {
  tenantId: number;
  slug: string;
  city: string;
  name: string;
  totalXP: number;
  rank: number;
}

export interface LeaderboardResponse {
  leaderboard: LeaderboardEntry[];
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  icon: string;
  unlockHint: string;
  earned: boolean;
  earnedAt?: string | null;
}

export interface BadgesResponse {
  badges: Badge[];
}

export interface ChallengeProgress {
  id: number;
  challengeKey: string;
  title: string;
  description: string;
  targetValue: number;
  xpReward: number;
  currentValue: number;
  completedAt?: string | null;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastActiveWeek?: string | null;
}

export interface GamificationMeResponse {
  totalXP: number;
  rank: number;
  totalCities: number;
  streak: StreakInfo;
  challenges: ChallengeProgress[];
  badges: Badge[];
}

export type GetTenantConfigParams = {
  /**
   * The tenant's subdomain slug (e.g. "austin")
   */
  slug: string;
};