# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Raj's Austin Events newsletter
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/digest/latest:
    get:
      operationId: getLatestDigest
      tags: [events]
      summary: Get the latest weekly digest
      responses:
        "200":
          description: Latest digest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigestResponse"
        "404":
          description: No digest found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/digest/generate:
    post:
      operationId: generateDigest
      tags: [events]
      summary: Generate a new weekly digest (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDigestRequest"
      responses:
        "200":
          description: Digest generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigestResponse"
  /events/digest/send:
    post:
      operationId: sendDigest
      tags: [events]
      summary: Send the latest digest to all subscribers (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDigestRequest"
      responses:
        "200":
          description: Digest sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /gamification/leaderboard:
    get:
      operationId: getGamificationLeaderboard
      tags: [gamification]
      summary: Get leaderboard of all cities ranked by XP (admin)
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gamification/me:
    get:
      operationId: getGamificationMe
      tags: [gamification]
      summary: Get current tenant's XP, rank, badges, challenges, and streak (admin)
      responses:
        "200":
          description: Gamification profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamificationMeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gamification/badges:
    get:
      operationId: getGamificationBadges
      tags: [gamification]
      summary: Get all badge definitions with earned state for current tenant (admin)
      responses:
        "200":
          description: Badges list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    EventItem:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        venue:
          type: string
        description:
          type: string
        link:
          type: string
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        featured:
          type: boolean
      required:
        - title
        - date
        - venue
        - description
        - category
    Digest:
      type: object
      properties:
        id:
          type: integer
        weekOf:
          type: string
          format: date-time
        subject:
          type: string
        intro:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventItem"
        sentAt:
          type: string
          format: date-time
          nullable: true
        sentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weekOf
        - subject
        - intro
        - events
        - sentCount
        - createdAt
    DigestResponse:
      type: object
      properties:
        digest:
          $ref: "#/components/schemas/Digest"
      required:
        - digest
...
    GenerateDigestRequest:
      type: object
      properties:
        weekOf:
          type: string
          description: ISO date string for the week (defaults to next Sunday)
        customNotes:
          type: string
          description: Optional personal notes to include in the intro
      required: []
    SendDigestRequest:
      type: object
      properties:
        digestId:
          type: integer
        testEmail:
          type: string
          format: email
          description: If provided, only sends to this test email
      required:
        - digestId
...
    LeaderboardEntry:
      type: object
      properties:
        tenantId:
          type: integer
        slug:
          type: string
        city:
          type: string
        name:
          type: string
        totalXP:
          type: integer
        rank:
          type: integer
      required:
        - tenantId
        - slug
        - city
        - name
        - totalXP
        - rank
    LeaderboardResponse:
      type: object
      properties:
        leaderboard:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
      required:
        - leaderboard
    Badge:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockHint:
          type: string
        earned:
          type: boolean
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - description
        - icon
        - unlockHint
        - earned
    BadgesResponse:
      type: object
      properties:
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
      required:
        - badges
    ChallengeProgress:
      type: object
      properties:
        id:
          type: integer
        challengeKey:
          type: string
        title:
          type: string
        description:
          type: string
        targetValue:
          type: integer
        xpReward:
          type: integer
        currentValue:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - challengeKey
        - title
        - description
        - targetValue
        - xpReward
        - currentValue
    StreakInfo:
      type: object
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastActiveWeek:
          type: string
          nullable: true
      required:
        - currentStreak
        - longestStreak
    GamificationMeResponse:
      type: object
      properties:
        totalXP:
          type: integer
        rank:
          type: integer
        totalCities:
          type: integer
        streak:
          $ref: "#/components/schemas/StreakInfo"
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeProgress"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
      required:
        - totalXP
        - rank
        - totalCities
        - streak
        - challenges
        - badges