# lib/api-zod/src/generated/types/badge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */

export interface Badge {
  id: string;
  name: string;
  description: string;
  icon: string;
  unlockHint: string;
  earned: boolean;
  earnedAt?: Date | null;
}