# lib/api-zod/src/generated/types/badgesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from "./badge";

export interface BadgesResponse {
  badges: Badge[];
}