# lib/api-zod/src/generated/types/challengeProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeProgress {
  id: number;
  challengeKey: string;
  title: string;
  description: string;
  targetValue: number;
  xpReward: number;
  currentValue: number;
  completedAt?: Date | null;
}