# lib/api-zod/src/generated/types/gamificationMeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from "./badge";
import type { ChallengeProgress } from "./challengeProgress";
import type { StreakInfo } from "./streakInfo";

export interface GamificationMeResponse {
  totalXP: number;
  rank: number;
  totalCities: number;
  streak: StreakInfo;
  challenges: ChallengeProgress[];
  badges: Badge[];
}