# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Raj's Austin Events newsletter
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  tenantId: number;
  slug: string;
  city: string;
  name: string;
  totalXP: number;
  rank: number;
}